+++ frontend/src/components/models/employee/EmployeePopover.tsx
@@ -6,7 +6,7 @@ import { Params } from 'types/status'
 import { theme } from 'styles/theme'
 import { convertIntoRGBA } from 'utils/color/convertIntoRGBA'
 import { calculateMinSizeBasedOnFigmaWidth } from 'utils/calculateSizeBasedOnFigma'
-import { BasicPopover } from 'components/ui/modal/BasicPopover'
+import { BasicPopover } from 'components/ui/popup/BasicPopover'
 import { EmployeeStatus } from 'components/models/employee/EmployeeStatus'
 import { Avatar } from 'components/ui/item/Avatar'
 import { Level } from 'components/ui/item/Level'
